pyproject.toml
[project]
name = "202510_learning"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "gputil>=1.4.0",
    "jupyter>=1.1.1",
    "lmdb>=1.7.3",
    "matplotlib>=3.10.7",
    "numpy>=2.2.6",
    "opencv-python>=4.11.0.86",
    "pandas>=2.3.3",
    "pillow>=11.3.0",
    "scikit-learn>=1.7.2",
    "tensorboard>=2.20.0",
    "timm>=1.0.20",
    "torch>=2.6.0",
    "torchaudio>=2.6.0",
    "torchsummary>=1.5.1",
    "torchvision>=0.21.0",
]

[dependency-groups]
dev = []

[tool.uv.sources]
torch = [
    { index = "pytorch-cu126", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
torchvision = [
    { index = "pytorch-cu126", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]

[[tool.uv.index]]
name = "pytorch-cu126"
url = "https://download.pytorch.org/whl/cu126"
explicit = true
